requirements.txt
django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
django-phonenumber-field==7.0.2
phonenumberslite==8.13.7
rest-framework-simplejwt==0.0.2


